# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Campaign Service",
    "description": "API for managing marketing campaigns",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3003"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service API key (CAMPAIGN_SERVICE_API_KEY)"
      }
    },
    "schemas": {
      "Campaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orgId": {
            "type": "string"
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "workflowSlug": {
            "type": "string"
          },
          "workflowDynastySlug": {
            "type": "string",
            "nullable": true
          },
          "featureDynastySlug": {
            "type": "string",
            "nullable": true
          },
          "brandUrl": {
            "type": "string",
            "nullable": true
          },
          "brandIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "featureSlug": {
            "type": "string",
            "nullable": true
          },
          "featureInputs": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "nullable": true
            }
          },
          "maxBudgetDailyUsd": {
            "type": "string",
            "nullable": true
          },
          "maxBudgetWeeklyUsd": {
            "type": "string",
            "nullable": true
          },
          "maxBudgetMonthlyUsd": {
            "type": "string",
            "nullable": true
          },
          "maxBudgetTotalUsd": {
            "type": "string",
            "nullable": true
          },
          "maxLeads": {
            "type": "integer",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "toResumeAt": {
            "type": "string",
            "nullable": true
          },
          "notifyFrequency": {
            "type": "string",
            "nullable": true
          },
          "notifyChannel": {
            "type": "string",
            "nullable": true
          },
          "notifyDestination": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orgId",
          "createdByUserId",
          "name",
          "workflowSlug",
          "workflowDynastySlug",
          "featureDynastySlug",
          "brandUrl",
          "brandIds",
          "featureSlug",
          "featureInputs",
          "maxBudgetDailyUsd",
          "maxBudgetWeeklyUsd",
          "maxBudgetMonthlyUsd",
          "maxBudgetTotalUsd",
          "maxLeads",
          "startDate",
          "endDate",
          "status",
          "toResumeAt",
          "notifyFrequency",
          "notifyChannel",
          "notifyDestination",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateCampaignBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "workflowSlug": {
            "type": "string",
            "minLength": 1
          },
          "workflowDynastySlug": {
            "type": "string",
            "minLength": 1
          },
          "orgId": {
            "type": "string",
            "minLength": 1
          },
          "brandUrl": {
            "type": "string",
            "minLength": 1
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1
          },
          "featureSlug": {
            "type": "string",
            "minLength": 1
          },
          "featureDynastySlug": {
            "type": "string",
            "minLength": 1
          },
          "featureInputs": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "maxBudgetDailyUsd": {
            "type": "string"
          },
          "maxBudgetWeeklyUsd": {
            "type": "string"
          },
          "maxBudgetMonthlyUsd": {
            "type": "string"
          },
          "maxBudgetTotalUsd": {
            "type": "string"
          },
          "maxLeads": {
            "type": "integer"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "notifyFrequency": {
            "type": "string"
          },
          "notifyChannel": {
            "type": "string"
          },
          "notifyDestination": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "orgId",
          "brandUrl",
          "brandIds"
        ]
      },
      "UpdateCampaignBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brandUrl": {
            "type": "string"
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "featureSlug": {
            "type": "string",
            "minLength": 1
          },
          "featureDynastySlug": {
            "type": "string",
            "minLength": 1
          },
          "featureInputs": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "maxBudgetDailyUsd": {
            "type": "string"
          },
          "maxBudgetWeeklyUsd": {
            "type": "string"
          },
          "maxBudgetMonthlyUsd": {
            "type": "string"
          },
          "maxBudgetTotalUsd": {
            "type": "string"
          },
          "maxLeads": {
            "type": "integer"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "activate",
              "stop"
            ]
          },
          "notifyFrequency": {
            "type": "string"
          },
          "notifyChannel": {
            "type": "string"
          },
          "notifyDestination": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List campaigns for org",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "featureDynastySlug",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaigns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Campaign"
                      }
                    }
                  },
                  "required": [
                    "campaigns"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create a new campaign",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaign": {
                      "$ref": "#/components/schemas/Campaign"
                    }
                  },
                  "required": [
                    "campaign"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get a specific campaign",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaign": {
                      "$ref": "#/components/schemas/Campaign"
                    }
                  },
                  "required": [
                    "campaign"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Update a campaign",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaign": {
                      "$ref": "#/components/schemas/Campaign"
                    }
                  },
                  "required": [
                    "campaign"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete a campaign",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/campaigns/list": {
      "get": {
        "tags": [
          "Scheduler"
        ],
        "summary": "List all campaigns across all orgs",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All campaigns with org info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaigns": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Campaign"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "externalOrgId": {
                                "type": "string"
                              },
                              "brandDomain": {
                                "type": "string",
                                "nullable": true
                              },
                              "brandName": {
                                "type": "string",
                                "nullable": true
                              }
                            },
                            "required": [
                              "externalOrgId",
                              "brandDomain",
                              "brandName"
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "campaigns"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}